add transition of push/replace scene, add MyButton Class

+++ Classes/Scene/FarmScene.cpp
@@ -66,6 +66,8 @@ bool Farm::init() {
 	Bag* bag = Bag::getInstance();
 	this->addChild(bag, 4);
 
+	// auto exitButton = ui::Button::create()
+
 	createFestivals();
 
 	// 初始化键盘监听器
@@ -168,7 +170,7 @@ void Farm::showDialogueOptions(Npc* npc, Sprite* dialogueBackground, Label* labe
 	const float optionY = dialogueBackground->getPositionY() + 120;  // 选项显示位置
 
 	// 创建选项按钮
-	for (int i = 0; i < options.size(); ++i) {
+	for (size_t i = 0; i < options.size(); ++i) {
 		auto optionButton = ui::Button::create();  // 创建按钮
 		optionButton->setTitleText(options[i]);
 		optionButton->setTitleFontSize(30);
